refactor: Move unit tests to test/unit
