+++ packages/fury-adapter-oas3-parser/test/mocha.opts
@@ -1,0 +1,1 @@
+--recursive
