+++ packages/fury-adapter-oas3-parser/test/unit/parser/oas/test_parseParameterObject.js
@@ -1,6 +1,6 @@
 const { expect } = require('chai');
 const { Fury } = require('fury');
-const parse = require('../../../lib/parser/oas/parseParameterObject');
+const parse = require('../../../../lib/parser/oas/parseParameterObject');
 
 const { minim } = new Fury();
